Issue in Inventory Loss Improvement Solved

+++ Project/medisync-backend/src/main/java/com/medisync/medisync_backend/dto/InventoryLossRequest.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class InventoryLossRequest {
 
-    private Integer stockId;
+    private Long stockId;
 
     private Integer quantity;
